refactor: coloca a pasta de columns dentro de CoeqDataTable no padrão da codebase
