--- src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/CoeqDataTable/consts.ts
@@ -1,25 +1,0 @@
-type IconDetails = {
-    src: string;
-    alt: string;
-};
-
-export type TagIconDetailsMap = Record<string, IconDetails>;
-
-export const iconDetailsMap: TagIconDetailsMap = {
-    danger: {
-        src: "https://media.graphassets.com/TWH6Oby6QuTFyq0wH9QK",
-        alt: "Ícone com símbolo da letra x",
-    },
-    warning: {
-        src: "https://media.graphassets.com/o0OkjNboRCqy2bYrRNnb",
-        alt: "Ícone de uma exclamação",
-    },
-    success: {
-        src: "https://media.graphassets.com/4qKuRCxHSySL23zxLd9b",
-        alt: "Ícone de uma marca de verificação",
-    },
-    pending: {
-        src: "https://media.graphassets.com/IdqIxy4LQAeIZfe9hWZK",
-        alt: "Ícone de uma ampulheta",
-    },
-};
