+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/CoeqDataTable/consts.tsx
@@ -1,10 +1,36 @@
-import { RenderStatusTagCell } from "./formatters/columns";
 import type { GridColDef } from "@mui/x-data-grid";
-import { DateRenderCell, CpfOrBirthdayFormatter } from "./formatters/columns";
-import { NameFormatter } from "./formatters/name";
-//TODO investigar como a conversao para string no dataGrid
+import { CpfOrBirthdayFormatter } from "./modules/CpfOrBirthdayFormatter";
+import { DateRenderCell } from "./modules/DateRenderCell";
+import { NameFormatter } from "./modules/NameFormatter";
+import { RenderStatusTagCell } from "./modules/RenderStatusTagCell";
 
-export const diabetesColumns: Array<GridColDef> = [
+type IconDetails = {
+    src: string;
+    alt: string;
+};
+
+export type TagIconDetailsMap = Record<string, IconDetails>;
+
+export const iconDetailsMap: TagIconDetailsMap = {
+    danger: {
+        src: "https://media.graphassets.com/TWH6Oby6QuTFyq0wH9QK",
+        alt: "Ícone com símbolo da letra x",
+    },
+    warning: {
+        src: "https://media.graphassets.com/o0OkjNboRCqy2bYrRNnb",
+        alt: "Ícone de uma exclamação",
+    },
+    success: {
+        src: "https://media.graphassets.com/4qKuRCxHSySL23zxLd9b",
+        alt: "Ícone de uma marca de verificação",
+    },
+    pending: {
+        src: "https://media.graphassets.com/IdqIxy4LQAeIZfe9hWZK",
+        alt: "Ícone de uma ampulheta",
+    },
+};
+
+export const coeqColumns: Array<GridColDef> = [
     {
         field: "patientName",
         headerName: "Nome",
