+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/CoeqDataTable/modules/DateRenderCell.tsx
@@ -1,0 +1,17 @@
+import { formatDate } from "@/features/common/time";
+import React from "react";
+import { RenderDateTagCell } from "./RenderDateTagCell";
+
+type DateRenderCellProps = {
+    value: string | null;
+};
+
+export const DateRenderCell: React.FC<DateRenderCellProps> = ({ value }) => {
+    if (!value)
+        return (
+            <div data-testid="tag">
+                <RenderDateTagCell />
+            </div>
+        );
+    return <div>{formatDate(new Date(value.slice(0, 10)))}</div>;
+};
