+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/CoeqDataTable/modules/NameFormatter.tsx
@@ -1,4 +1,4 @@
-import { nameFormatter } from "../../../../../../../logic";
+import { nameFormatter } from "../../../../../logic"; //TODO: Checar este import
 
 type NameFormatterProps = {
     value: string | null;
