+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/CoeqDataTable/modules/RenderDateTagCell.tsx
@@ -1,0 +1,11 @@
+import { TableTag } from "@/features/acf/common/frontend/TableTag";
+import React from "react";
+import { iconDetailsMap } from "../consts";
+
+export const RenderDateTagCell: React.FC = () => (
+    <TableTag
+        theme="pending"
+        text="Não realizada"
+        icon={iconDetailsMap.pending}
+    />
+);
